# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MindWeaver AI",
        "title": "自动化自媒体工具-API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ctsch",
            "url": "http://www.example.com/support",
            "email": "xxxxxxx@xxx.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "192.168.0.106:14010",
    "basePath": "/api",
    "paths": {
...
        "/prompts/test": {
            "post": {
                "description": "Builds a system prompt using the given parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prompts"
                ],
                "summary": "Generate a system prompt using the prompt engine",
                "parameters": [
                    {
                        "description": "Request data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.TestPromptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.TestPromptResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.TestPromptRequest": {
            "type": "object",
            "properties": {
                "customModeConfigs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sections.ModeConfig"
                    }
                },
                "envContext": {
                    "$ref": "#/definitions/prompts.EnvironmentContext"
                },
                "globalInstructions": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "prompt": {
                    "type": "string"
                }
            }
        },
        "api.TestPromptResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        },
...
        "base.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "prompts.EnvironmentContext": {
            "type": "object",
            "properties": {
                "browserViewportSize": {
                    "description": "e.g., \"1280x800\"",
                    "type": "string"
                },
                "cwd": {
                    "type": "string"
                },
                "language": {
                    "description": "e.g., \"en\", \"fr\"",
                    "type": "string"
                },
                "supportsComputerUse": {
                    "description": "For browser actions etc.",
                    "type": "boolean"
                }
            }
        },
        "sections.FileRestriction": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "fileRegex": {
                    "description": "Go regex syntax",
                    "type": "string"
                }
            }
        },
        "sections.GroupEntry": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "The core group name",
                    "allOf": [
                        {
                            "$ref": "#/definitions/toolgroups.ToolGroupName"
                        }
                    ]
                },
                "restriction": {
                    "description": "Optional restriction",
                    "allOf": [
                        {
                            "$ref": "#/definitions/sections.FileRestriction"
                        }
                    ]
                }
            }
        },
        "sections.ModeConfig": {
            "type": "object",
            "properties": {
                "customInstructions": {
                    "type": "string"
                },
                "groups": {
                    "description": "Array of allowed tool groups/restrictions",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sections.GroupEntry"
                    }
                },
                "name": {
                    "type": "string"
                },
                "roleDefinition": {
                    "type": "string"
                },
                "slug": {
                    "$ref": "#/definitions/sections.ModeSlug"
                }
            }
        },
        "sections.ModeSlug": {
            "type": "string",
            "enum": [
                "code"
            ],
            "x-enum-varnames": [
                "DefaultModeSlug"
            ]
        },
...
        "toolgroups.ToolGroupName": {
            "type": "string",
            "enum": [
                "read",
                "edit",
                "command",
                "browser",
                "mcp"
            ],
            "x-enum-comments": {
                "GroupMcp": "Keep if MCP tools might be re-added"
            },
            "x-enum-varnames": [
                "GroupRead",
                "GroupEdit",
                "GroupCommand",
                "GroupBrowser",
                "GroupMcp"
            ]
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}